refactor: SignupRepository を UserRepository に統合 (#603)
1 Aggregate : 1 Repository の原則に沿い、User Aggregate に対する
重複リポジトリを解消。SignupRepository の emailExists / createUser を
UserRepository に統合し、関連する DI 配線・テスト・ドキュメントを更新。
