+++ server/domain/models/user/user-repository.ts
@@ -1,6 +1,12 @@
 import type { UserId } from "@/server/domain/common/ids";
 import type { User, ProfileVisibility } from "@/server/domain/models/user/user";
 
+export type SignupData = {
+  email: string;
+  password: string;
+  name: string | null;
+};
+
 export type UserRepository = {
   findById(id: UserId): Promise<User | null>;
   findByIds(ids: readonly UserId[]): Promise<User[]>;
@@ -11,7 +17,7 @@ export type UserRepository = {
     name: string | null,
     email: string | null,
   ): Promise<void>;
-  emailExists(email: string, excludeUserId: UserId): Promise<boolean>;
+  emailExists(email: string, excludeUserId?: UserId): Promise<boolean>;
   findPasswordHashById(id: UserId): Promise<string | null>;
   findPasswordChangedAt(id: UserId): Promise<Date | null>;
   updatePasswordHash(id: UserId, passwordHash: string): Promise<void>;
@@ -19,4 +25,5 @@ export type UserRepository = {
     id: UserId,
     visibility: ProfileVisibility,
   ): Promise<void>;
+  createUser(data: SignupData): Promise<UserId>;
 };
